iadded specific text in confirmation email for local groups assemblyes
* changed logo, body text and signature in local group assembly confirmation email
