+++ som_cas/templatetags/assembly_event.py
@@ -13,5 +13,5 @@ def assembly_event(value):
     conector_text = ''
     if value.local_group.name[0].upper() in ['AEIOU']:
         conector_text = _('l\'')
-    base_text = _("l'Assamblea del grup local de")
-    return f'{base_text} {conector_text}{value.local_group.name}'
+    base_text = _("l'Assemblea Ordinària de la")
+    return f'{base_text} {conector_text}{value.local_group.full_name}'
